+++ lib/src/model/answer/time_answer_format.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:survey_kit/src/model/answer/answer.dart';
 
-part 'time_answer_formart.g.dart';
+part 'time_answer_format.g.dart';
 
 @JsonSerializable()
 class TimeAnswerFormat implements Answer {