+++ lib/src/util/extension.dart
@@ -2,6 +2,6 @@ extension JoinedIterable<T> on Iterable<T> {
   Iterable<T> withSeparator(T separator) {
     return isEmpty
         ? this
-        : (expand((element) => [element, separator]).toList()..removeLast());
+        : (expand((element) => [element, separator]).toList());
   }
 }