+++ lib/src/view/answer/multiple_choice_auto_complete_answer_view.dart
@@ -6,8 +6,8 @@ import 'package:survey_kit/src/model/answer/option.dart';
 import 'package:survey_kit/src/model/result/step_result.dart';
 import 'package:survey_kit/src/model/step.dart';
 import 'package:survey_kit/src/util/measure_date_state_mixin.dart';
+import 'package:survey_kit/src/view/answer/selection_list_tile.dart';
 import 'package:survey_kit/src/view/content/content_widget.dart';
-import 'package:survey_kit/src/view/selection_list_tile.dart';
 import 'package:survey_kit/src/view/step_view.dart';
 
 class MultipleChoiceAutoCompleteAnswerView extends StatefulWidget {