+++ lib/src/view/answer/single_choice_answer_view.dart
@@ -4,8 +4,8 @@ import 'package:survey_kit/src/model/answer/single_select_answer.dart';
 import 'package:survey_kit/src/model/result/step_result.dart';
 import 'package:survey_kit/src/model/step.dart';
 import 'package:survey_kit/src/util/measure_date_state_mixin.dart';
+import 'package:survey_kit/src/view/answer/selection_list_tile.dart';
 import 'package:survey_kit/src/view/content/content_widget.dart';
-import 'package:survey_kit/src/view/selection_list_tile.dart';
 import 'package:survey_kit/src/view/step_view.dart';
 
 class SingleChoiceAnswerView extends StatefulWidget {