+++ lib/src/widget/survey_kit_markdown.dart
@@ -11,9 +11,12 @@ class SurveyKitMarkdown extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MarkdownBody(
-      data: markdown,
-      selectable: true,
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 8),
+      child: MarkdownBody(
+        data: markdown,
+        selectable: true,
+      ),
     );
   }
 }